requirements.txt
streamlit==1.39.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
numpy>=2.1.1
Pillow==10.3.0
